fix：修复会员性别 Signed-off-by: HaiFeng <huanghaifeng@regentsoft.net>

+++ rbp-api-core/src/main/java/com/regent/rbp/api/core/member/MemberCard.java
@@ -47,7 +47,7 @@ public class MemberCard {
     private String phone;
 
     @ApiModelProperty(notes = "性别")
-    private Long sexId;
+    private Integer sexId;
 
     @ApiModelProperty(notes = "来源类别")
     private Integer originType;
